HZulip

Documentation

data Message

Represents a Zulip Message

Constructors

Message

Fields

messageId :: String

messageType :: String

messageContent :: String

messageAvatarUrl :: String

messageTimestamp :: Int

messageDisplayRecipient :: Either String [User]

messageSender :: User

messageGravatarHash :: String

messageRecipientId :: String

messageClient :: String

messageSubjectLinks :: [String]

messageSubject :: String

data ZulipClient

Represents a Zulip API client

Constructors

ZulipClient

Fields

clientEmail :: String

clientApiKey :: String

clientBaseUrl :: String

defaultBaseUrl :: String

The default zulip API URL

getEvents :: ZulipClient -> Queue -> Bool -> IO [Event]

newZulip :: String -> String -> ZulipClient

Helper for creating a ZulipClient with the baseUrl set to defaultBaseUrl

registerQueue :: ZulipClient -> [String] -> Bool -> IO Queue

This registers a new event queue with the zulip API. It's a lower level function, which shouldn't be used unless you know what you're doing. It takes a ZulipClient, a list of names of the events you want to listen for and whether you'd like for the content to be rendered in HTML format (if you set the last parameter to False it will be kept as typed, in markdown format)

sendMessage :: ZulipClient -> String -> [String] -> String -> String -> IO String

This wraps `POST https:/api.zulip.comv1/messages` with a nicer root API. Simpler helpers for each specific case of this somewhat overloaded endpoint will also be provided in the future.

It takes the message mtype, mrecipients, msubject and mcontent and returns the created message's id in the IO monad.