HZulip

Documentation

data Event

Represents zulip events

Constructors

Event

Fields

eventType :: String

eventId :: Int

eventMessage :: Maybe Message

data Message

Represents a Zulip Message

Constructors

Message

Fields

messageId :: Int

messageType :: String

messageContent :: String

messageAvatarUrl :: String

messageTimestamp :: Int

messageSender :: User

messageGravatarHash :: String

messageRecipientId :: Int

messageClient :: String

messageSubjectLinks :: [String]

messageSubject :: String

data Queue

Represents some event queue

Constructors

Queue

Fields

queueId :: String

lastEventId :: Int

data User

Represents a zulip user account - for both display_recipient and message_sender representations

Constructors

User

Fields

userId :: Int

userFullName :: String

userEmail :: String

userDomain :: String

userShortName :: String

data ZulipClient

Represents a Zulip API client

Constructors

ZulipClient

Fields

clientEmail :: String

clientApiKey :: String

clientBaseUrl :: String

type EventCallback = Event -> IO ()

The root type for Event callbacks

type MessageCallback = Message -> IO ()

Type for message callbacks

defaultBaseUrl :: String

The default zulip API URL

eventTypes :: [String]

The list of all avaiable event types

getEvents :: ZulipClient -> Queue -> Bool -> IO (Queue, [Event])

Fetches new set of events from a Queue.

newZulip :: String -> String -> ZulipClient

Helper for creating a ZulipClient with the baseUrl set to defaultBaseUrl

onNewEvent :: ZulipClient -> [String] -> EventCallback -> IO ()

Registers an event callback for specified events and keeps executing it over events as they come in. Will loop forever

onNewMessage :: ZulipClient -> MessageCallback -> IO ()

Registers a callback to be executed whenever a message comes in. Will loop forever

registerQueue :: ZulipClient -> [String] -> Bool -> IO Queue

This registers a new event queue with the zulip API. It's a lower level function, which shouldn't be used unless you know what you're doing. It takes a ZulipClient, a list of names of the events you want to listen for and whether you'd like for the content to be rendered in HTML format (if you set the last parameter to False it will be kept as typed, in markdown format)

sendMessage :: ZulipClient -> String -> [String] -> String -> String -> IO Int

This wraps `POST https:/api.zulip.comv1/messages` with a nicer root API. Simpler helpers for each specific case of this somewhat overloaded endpoint will also be provided in the future.

It takes the message mtype, mrecipients, msubject and mcontent and returns the created message's id in the IO monad.

sendPrivateMessage :: ZulipClient -> [String] -> String -> IO Int

Helper for sending private messages. Takes the list of recipients and the message's content.

sendStreamMessage :: ZulipClient -> String -> String -> String -> IO Int

Helper for sending stream messages. Takes the stream name, the subject and the message.